# docs/swagger.yaml
definitions:
  handler.CreateTaskRequest:
    properties:
      description:
        type: string
      expires_at:
        type: string
      name:
        type: string
    required:
    - description
    - expires_at
    - name
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  handler.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  handler.TaskDetailsResponse:
    properties:
      data:
        properties:
          created_at:
            type: string
          description:
            type: string
          expires_at:
            type: string
          id:
            type: integer
          name:
            type: string
          updated_at:
            type: string
        type: object
      success:
        type: boolean
    type: object
  handler.TaskListResponse:
    properties:
      data:
        items:
          properties:
            description:
              type: string
            expires_at:
              type: string
            id:
              type: integer
            name:
              type: string
          type: object
        type: array
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/v1/task:
    get:
      consumes:
      - application/json
      description: Get a paginated list of tasks
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page_number
        required: true
        type: integer
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TaskListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task with the input payload
      parameters:
      - description: Task object that needs to be added
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new task
      tags:
      - tasks
  /api/v1/task/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a task by its ID
      parameters:
      - description: Task ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get a single task by its ID
      parameters:
      - description: Task ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TaskDetailsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update an existing task by ID with the input payload
      parameters:
      - description: Task ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Updated task object
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update a task
      tags:
      - tasks
swagger: "2.0"